templates/requirements.txt
Flask
pillow
pytesseract
transformers
torch
